api_gateway/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
requests
python-dotenv
